requirements.txt
pandas
fastapi
uvicorn
python-dotenv
tensorflow
numpy
opencv-python
matplotlib
scikit-learn